# api.yaml
openapi: 3.0.1
info:
  title: Base
  description: Description
  version: 1.0.0
servers:
  - url: 'https://api-ecommerce.hakaya.technology'
  - url: 'http://localhost:3000'
paths:
...
  /categories/{categoryId}/subs:
    post:
      tags:
        - Item
      summary: Add Category Item
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []

      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/subCategorySchema'
      responses:
        200:
          $ref: '#/components/responses/item'
    get:
      tags:
        - Item
      summary: get category items
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string

      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/item'
  /categories/{categoryId}/subs/{subId}:
    put:
      tags:
        - Item
      summary: update Category Item
      security:
        - ApiKeyAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string

        - name: subId
          in: path
          required: true
          schema:
            type: string

      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/subCategorySchema'
      responses:
        200:
          $ref: '#/components/responses/item'

    get:
      tags:
        - Item
      summary: Get Category Item
      parameters:
        - name: categoryItem
          in: path
          required: true
      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/item'

    delete:
      tags:
        - Item
      summary: Delete Category Item
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
        - name: subId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/message'
...
  /orders:
    post:
      tags:
        - Order
      summary: make Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/oneOrderSchema'
      responses:
        200:
          $ref: '#/components/responses/ordersRes'
    get:
      tags:
        - Order
      summary: get orders
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          $ref: '#/components/responses/ordersRes'
  /orders/{orderId}:
    put:
      tags:
        - Order
      summary: update Order status
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/orderStatusSchema'
      responses:
        200:
          $ref: '#/components/responses/ordersRes'
    delete:
      tags:
        - Order
      summary: delete Order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          $ref: '#/components/responses/message'
...
  /users:
    get:
      tags:
        - User
      summary: get all users
      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/user'
  /users/{userId}:
    get:
      tags:
        - User
      summary: get user details
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/user'
    put:
      tags:
        - User
      summary: update user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/updateUser'
      responses:
        200:
          $ref: '#/components/responses/user'
  /users/{userId}/orders:
    get:
      tags:
        - User
      summary: get my Orders ( you can filter by sending query param status=current or previous)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/ordersRes'
  /users/{userId}/addresses:
    post:
      tags:
        - User
      summary: add address user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addressSchema'
      responses:
        200:
          $ref: '#/components/responses/user'
  /users/{userId}/addresses/{addressId}:
    put:
      tags:
        - User
      summary: add address user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: addressId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addressSchema'
      responses:
        200:
          $ref: '#/components/responses/user'
    delete:
      tags:
        - User
      summary: update user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: addressId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/user'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  responses:
...
    user:
      description: Get All Users.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserSchema'
...
    ordersRes:
      description: get search.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ordersSchema'
...
    message:
      description: Message.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/messageSchema'
...
  schemas:
...
    messageSchema:
      type: object
      properties:
        message:
          type: "string"
          example: "Item has been deleted"
...
    subCategorySchema:
      type: object
      properties:
        data:
          type: object
          properties:
            name:
              type: "string"
              example: "Item1"
            category:
              type: "string"
            __v:
              type: "number"
              example: 0
        logo:
          type: string
          format: binary
...
    UserSchema:
      type: object
      properties:
        _id:
          type: "string"
        mobile:
          type: "string"
        name:
          type: "string"
        email:
          type: "string"
        fcm_token:
          type: "string"
        roles:
          type: "array"
          items:
            type: "string"
        profile_image:
          type: "string"
        device_id:
          type: "string"
        search_inputs:
          type: "array"
          items:
            type: "string"
        addresses:
          type: "array"
          items:
            type: "object"
            properties:
              name:
                type: "string"
              address:
                type: "string"
              mobile:
                type: "string"
              special_mark:
                type: "string"
              city:
                type: "string"
              area:
                type: "string"
              default:
                type: "boolean"
    updateUser:
      type: object
      properties:
        data:
          type: object
          properties:
            name:
              type: "string"
            mobile:
              type: "string"
            email:
              type: "string"
        profile_image:
          type: string
          format: binary
    addressSchema:
      type: "object"
      properties:
        address:
          type: "string"
        name:
          type: "string"
        mobile:
          type: "string"
        special_mark:
          type: "string"
        city:
          type: "string"
        area:
          type: "string"
        default:
          type: "boolean"
...
    oneOrderSchema:
      properties:
        products:
          type: "array"
          items:
            type: "object"
            properties:
              product_id:
                type: "string"
              count:
                type: "number"
              options:
                type: "array"
                items:
                  type: "object"
                  properties:
                    option_id:
                      type: "string"
                    is_selected:
                      type: "boolean"
                    choices:
                      type: "array"
                      items:
                        type: "string"
        promo_code:
          type: "string"
        total_price:
          type: "number"
        payment_type:
          type: "number"
        shipping_type:
          type: "number"
        address:
          type: "object"
          properties:
            name:
              type: "string"
            mobile:
              type: "string"
            city:
              type: "string"
            area:
              type: "string"
            address:
              type: "string"
            specila_mark:
              type: "string"
...
    ordersSchema:
      type: "object"
      properties:
        orders:
          type: "array"
          items:
            type: "object"
            properties:
              _id:
                type: "string"
              products:
                type: "array"
                items:
                  type: "object"
                  properties:
                    options:
                      type: "array"
                      items:
                        type: "undefined"
                    price_info:
                      required:
                        - "count"
                        - "product_price"
                        - "total_price"
                      properties:
                        count:
                          type: "number"
                        product_price:
                          type: "number"
                        total_price:
                          type: "number"
                      type: "object"
                    product:
                      properties:
                        name:
                          type: "string"
                        price:
                          type: "number"
                        main_image:
                          type: "string"
                        _id:
                          type: "string"
                        free_shipping:
                          type: "boolean"
                        subCategory_discount:
                          properties:
                            _id:
                              type: "string"
                            isArchived:
                              type: "boolean"
                            type:
                              type: "string"
                            value:
                              type: "number"
                            subCategory:
                              type: "string"
                            date_from:
                              type: "string"
                            date_to:
                              type: "string"
                            created_at:
                              type: "string"
                            updated_at:
                              type: "string"
                            __v:
                              type: "number"
                          type: "object"
                      type: "object"
              order_price_info:
                properties:
                  total_order_price:
                    type: "number"
                  discount:
                    type: "number"
                  tax:
                    type: "number"
                  tax_value:
                    type: "number"
                  delivery_cost:
                    type: "number"
                  additional_cost:
                    type: "number"
                  final_price:
                    type: "number"
                type: "object"
              address:
                properties:
                  city:
                    type: "string"
                  area:
                    type: "string"
                  address:
                    type: "string"
                  special_mark:
                    type: "string"
                  mobile:
                    type: "string"
                type: "object"
              user:
                required:
                  - "_id"
                  - "mobile"
                properties:
                  _id:
                    type: "string"
                  mobile:
                    type: "string"
                type: "object"
              status:
                type: "array"
                items:
                  type: "object"
                  properties:
                    text:
                      type: "string"
                    created_at:
                      type: "string"
              payment_type:
                type: "string"
              shipping_type:
                type: "string"
              created_at:
                type: "string"
              updated_at:
                type: "string"
              order_number:
                type: "string"
              __v:
                type: "number"
        pages:
          type: "number"
        length:
          type: "number"
    orderStatusSchema:
      type: "object"
      properties:
        status:
          type: "string"
          example:
            "completed"
...